yb-voyager-agent/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for MCP server communication
httpx==0.25.2
aiohttp==3.9.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment management
python-dotenv==1.0.0

# Google Cloud and Vertex AI
google-auth==2.23.4
google-cloud-aiplatform==1.38.1
google-cloud-core==2.3.3 
